videocontroller, gymcontroller

+++ database/migrations/2020_03_24_221215_create_gym_table.php
@@ -23,7 +23,7 @@ class CreateGymTable extends Migration
             $table->string('country');
             $table->string('zip_code');
             $table->string('website');
-            $table->unsignedBigInteger('owner_id');
+            $table->unsignedInteger('owner_id');
             $table->foreign('owner_id')->references('id')->on('users');
             $table->timestamps();
         });
